refactor:Changed the tab switching function to a module.
